Refactor name according to new naming convention
