+++ src/components/ChatCard.js
@@ -1,30 +1,49 @@
-import { Fragment } from 'react';
+import { Fragment } from "react";
+
+import useUserStore from "../stores/UserStore";
 
 export default function ChatCard({
-  chatPartnerID,
-  chatPartnerUsername,
-  chatDate,
-  chatPreviewText,
-  onClick, // Add this line to include the onClick prop
+  messageSenderNo,
+  messageRecieverNo,
+  messageDate,
+  messageText,
+  messageImagePointer,
 }) {
+  const {user} = useUserStore();
+
+  async function getMessageData() {}
+  const imageUrl = messageImagePointer
+    ? messageImagePointer.get("Image_File").url()
+    : null;
+
+
   return (
     <Fragment>
-      <div className="message-card-container" id={`chat-${chatPartnerID}`} onClick={onClick}>
-        {/* The rest of your ChatCard contents */}
-        <div className="card">
-          <div className="in-line">
-            <section className="message-card-body">
-              <div className="in-line">
-                <p className="bold" id={`newsTitle username-${chatPartnerID}`}>{chatPartnerUsername}</p>
-                <h5 id={`newsDate date-${chatPartnerID}`}>{chatDate}</h5>
-              </div>
-              <div className="meta-text">
-                <p>{chatPreviewText}</p>
-              </div>
-            </section>
+      <div
+        className="chat-card chats"
+        id={messageSenderNo + messageRecieverNo}
+      >
+        <section className="chat-card-body">
+          <div className="in-line-message">
+            <p className="bold" id="messageSender">
+              {"Sender: " + messageSenderNo + " Receiver: " + messageRecieverNo}
+            </p>
+            <h5 id="messageDate">{messageDate}</h5>
           </div>
-        </div>
+          <div className="meta-text">
+            <p>{messageText}</p>
+            {/* for rendering the image if there is one */}
+            {imageUrl && (
+              <img
+                src={imageUrl}
+                alt="Message Attachment"
+                className="message-image"
+              />
+            )}
+          </div>
+        </section>
       </div>
     </Fragment>
   );
+  
 }
