+++ src/components/ChatListHome.js
@@ -1,8 +1,11 @@
 import React, { Fragment } from 'react';
 import { Link } from "react-router-dom";
+
+//css import
 import "../../src/styles.css";
 
-import ChatListInbox from './ChatListInbox';
+//components import
+import ChatListInbox from './ChatListCardContainer';
 
 export default function ChatListHome() {
